# Conflicts:
#	dspace-api/src/main/java/org/dspace/app/ldn/service/impl/LDNMessageServiceImpl.java
#	dspace-server-webapp/src/test/java/org/dspace/app/rest/LDNInboxControllerIT.java

+++ dspace-api/src/main/java/org/dspace/qaevent/service/impl/QAEventActionServiceImpl.java
@@ -78,6 +78,12 @@ public class QAEventActionServiceImpl implements QAEventActionService {
             if (qaevent.getRelated() != null) {
                 related = itemService.find(context, UUID.fromString(qaevent.getRelated()));
             }
+            if (topicsToActions.get(qaevent.getTopic()) == null) {
+                String msg = "Unable to manage QA Event typed " + qaevent.getTopic()
+                    + ". Managed types are: " + topicsToActions;
+                log.error(msg);
+                throw new RuntimeException(msg);
+            }
             topicsToActions.get(qaevent.getTopic()).applyCorrection(context, item, related,
                 jsonMapper.readValue(qaevent.getMessage(), qaevent.getMessageDtoClass()));
             qaEventService.deleteEventByEventId(qaevent.getEventId());
